+++ Sources/SwiftKit/Authentication/LocalAuthenticationService.swift
@@ -76,4 +76,3 @@ open class LocalAuthenticationService: AuthenticationService {
     }
 }
 #endif
-