+++ Sources/SwiftKit/Concurrency/Collection+Async.swift
@@ -8,7 +8,6 @@
 
 import Foundation
 
-@available(iOS 15.0, macOS 12.0, watchOS 8.0, tvOS 15.0, *)
 public extension Collection {
     
     /**